requirements.txt
subprocess.run
ffmpeg
opencv-python
pillow
scikit-image
scikit-learn
matplotlib
pandas
ipykernel
jupyter
av
tensorflow